+++ src/Vehicles/Boat.java
@@ -76,7 +76,7 @@ public class Boat extends TransportVehicle {
     }
 
     @Override
-    public void update(float gameTime) throws Exception {
+    public void update(int gameTime) throws Exception {
         if (position == destination.getPosition()){
             // send message
             // wait for message depart
