+++ src/Vehicles/Train.java
@@ -60,7 +60,7 @@ public class Train extends TransportVehicle{
     }
 
     @Override
-    public void update(float gameTime) throws Exception {
+    public void update(int gameTime) throws Exception {
         if (position == destination.getPosition()){
             // send message
             // wait for message depart
